Code review stuff

+++ src/bindings/python/tests/test_ngraph/test_create_op.py
@@ -2106,12 +2106,12 @@ def test_irdft():
 
 
 def test_grid_sample():
-    img = ov.parameter([1, 3, 100, 100], name="image")
-    grid = ov.parameter([1, 10, 10, 2], name="grid")
+    img = ov.parameter([1, 3, 100, 100], dtype=np.int32, name="image")
+    grid = ov.parameter([1, 10, 10, 2], dtype=np.float32, name="grid")
 
     node = ov.grid_sample(img, grid, {})
 
     assert node.get_type_name() == "GridSample"
     assert node.get_output_size() == 1
     assert list(node.get_output_shape(0)) == [1, 3, 10, 10]
-    assert node.get_output_element_type(0) == Type.f32
+    assert node.get_output_element_type(0) == Type.i32
